https://yt.iteco.dev/issue/ISPP-1031

+++ processor/src/core/java/ru/axetta/ecafe/processor/core/partner/etpmv/ETPMVService.java
@@ -176,7 +176,7 @@ public class ETPMVService {
                 return;
             }
         }
-        List<Long> _benefits;
+        List<Integer> _benefits;
         if (newFormat) {
             _benefits = RuntimeContext.getAppContext().getBean(ETPMVDaoService.class).getDSZNBenefits(benefits);
         } else {
