+++ processor/src/web/java/ru/axetta/ecafe/processor/web/partner/integra/soap/ClientRoomControllerWS.java
@@ -10861,7 +10861,7 @@ public class ClientRoomControllerWS extends HttpServlet implements ClientRoomCon
                 throw new ClientNotFoundException(String.format("Unable to find client with guid={%s}", clientGuid));
             }
 
-            DAOUtils.createApplicationForFood(persistenceSession, client, Arrays.asList(otherDiscount ? null : categoryDiscount),
+            DAOUtils.createApplicationForFood(persistenceSession, client, Arrays.asList(otherDiscount ? null : categoryDiscount.intValue()),
                     mobilePhone, guardianName, guardianSecondName, guardianSurname, serviceNumber,
                     ApplicationForFoodCreatorType.PORTAL);
             DAOUtils.updateApplicationForFood(persistenceSession, client,
